requirements.txt
# Dev requirements file for local env, do not use in glitch.
APScheduler==3.10.1
black==23.7.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
Flask==2.3.2
Flask-APScheduler==1.12.4
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
platformdirs==3.8.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
waitress==2.1.2
Werkzeug==2.3.6
